Build NEMoE object from a phyloseq object.

Build NEMoE object from a phyloseq object.

NEMoE_buildFromPhyloseq(
  ps,
  Nutrition,
  Response,
  K = NULL,
  gatherParam = list(),
  filtParam = list(),
  transParam = list(),
  cvParams = list(),
  taxLevel = NULL,
  taxTab = NULL,
  standardize = TRUE,
  ...
)

Arguments

ps
A phyloseq object of input microbiome data

Nutrition
A dataframe of matrix of Nutrition matrix.

Response
A numeric(factor) vector of Health outcome.

K
A integer of number of latent class.

gatherParam
A list of parameters for gather phyloseq object. See psGather.

filtParam
A list of parameters for filtering of Microbiome data. can be set by prev (The number of non-zero proportion of features) and var(varaince of features).

transParam
A list of parameters for transformation of Microbiome data. See compTransform.

cvParams
A list of cross validation parameters.

taxLevel
A character of selected name of taxonomic levels.

taxTab
A dataframe of taxonomic table.

standardize
Logical flag for x variable standardization. Default is standardize=TRUE.

...
Other parameters can pass to NEMoE_buildFromPhyloseq. See createParameterList

Value

A NEMoE object.

Examples

data(PD)
ps = PD$ps
Response <- PD$data_list$Response
Nutrition_data <- PD$data_list$Nutrition
NEMoE <- NEMoE_buildFromPhyloseq(ps = ps, Nutrition = Nutrition_data, Response = Response)

#> 载入需要的程辑包：phyloseq