Create list of parameters for fitting NEMoE

This function create parameters that put into fitNEMoE function.

createParameterList(
  lambda1 = 0.02,
  lambda2 = 0.015,
  alpha1 = 0.5,
  alpha2 = 0.5,
  beta_max = 10,
  EM_alg = "EM",
  init = "kmeans",
  itmax = 100,
  itmin = 1,
  adapt = TRUE,
  btr = TRUE,
  stop_all = TRUE,
  verbose = TRUE,
  early_stop = FALSE
)

Arguments

lambda1
Penalty regularizer for the experts.

lambda2
Penalty regularizer for the gating network.

alpha1
Elastic net penalty value for experts.

alpha2
Elastic net penalty value for gating network.

beta_max
Maximal of coefficient. By default is 10.

EM_alg
Method for Expecation maximization update. Can be chosen from "EM", "CEM", "GEM", "SEM", "SAEM", "GEM". By default is "EM".

init
Method for initialization. Can be chosen from "rand", "kmeans" and "glmnet". If init="rand" will use a dirichlet distribution initialing the latent class. If init="kmeans" the latent class will initialized using kmeans clustering of input for gating network. If init="glmnet" will use a equal probability with lasso as its corresponding coefficients in experts network.

itmax
Maximium number of iteration in fitting NEMoE. By default is 100.

itmin
Minimium number of iteration in fitting NEMoE. By default is 3.

adapt
whether use adaptive mode in optimization. By default is TRUE.

btr
Whether use backtracking during the iteration. By default is TRUE.

stop_all
Method of stop criterion. If stop_all = TRUE means that either coefficient or loss function converge. If stop_all = FALSE means that both coefficient and loss function converge.

verbose
A logical input indicating whether the intermediate steps will be printed.

early_stop
A logical input indicate whether early stop when one of the fitted latent class have zero variables selected (to save time). By default is TRUE.

Value

A list contain parameters in fitting NEMoE.

See also

Examples

params = createParameterList(lambda1 = c(0.005, 0.01, 0.02, 0.025))