Fit NEMoE model

This function fit NEMoE model using EM algorithm.

fitNEMoE(
  NEMoE,
  beta_init = NULL,
  gamma_init = NULL,
  num_restart = 5,
  restart_it = 10,
  BPPARAM = BiocParallel::SerialParam(),
  ...
)

Arguments

NEMoE
a NEMoE object contain data and parameters for fitting.

beta_init
initial values of parameters in experts network.

gamma_init
initial value of parameters in gating network.

num_restart
Number of restart times.

restart_it
Maximial number of iterations during restart. the BiocParallel package is used. Default is SerialParam().

BPPARAM
A BiocParallelParam class object from.

...
other parameters can be passed to fitNEMoE. See createParameterList

Value

a NEMoE object contain fitted result including parameters beta a list contain coefficients of experts network in all levels. gamma a matrix of coefficients in gating network. r_i an array of inherit partition result in EM algorithm. pen_fac estimated penalty factor for lambda1 (adapt = T). The (balanced) lambda1 in NEMoE is lambda1*pen_fac latent estimated latent classes.

See also

Examples

data(NEMoE_example)
NEMoE_example

#> Numer of samples:  168 
#> Microbiome number of variables: 5, 16, 24, 36, 101, 
#> Nutrition dim: 168 27 
#> Number of latent class K:  2 
#> NEMoE output names (5): beta gamma r_i pen_fac LL
#> param names (14): lambda1 lambda2 ... early_stop init
#> cross validation param names (9): g1 lambda2_seq ... itmax_cv itmax_fit

NEMoE_example <- fitNEMoE(NEMoE_example, num_restart = 0)

#> Fitting NEMoE.... 
#> it: 0, PLL0:-1025.13
#> it: 1, PLL:-522.349
#> it: 2, PLL:-513.522
#> it: 3, PLL:-510.421
#> it: 4, PLL:-508.396
#> it: 5, PLL:-506.907
#> it: 6, PLL:-505.6
#> it: 7, PLL:-504.439
#> it: 8, PLL:-503.384
#> it: 9, PLL:-502.49
#> it: 10, PLL:-501.619
#> it: 11, PLL:-500.746
#> it: 12, PLL:-499.965
#> it: 13, PLL:-499.247
#> it: 14, PLL:-498.522
#> it: 15, PLL:-497.749
#> it: 16, PLL:-496.892
#> it: 17, PLL:-495.909
#> it: 18, PLL:-494.746
#> it: 19, PLL:-493.337
#> it: 20, PLL:-491.622
#> it: 21, PLL:-489.529
#> it: 22, PLL:-487.006
#> it: 23, PLL:-484.229
#> it: 24, PLL:-481.513
#> it: 25, PLL:-479.01
#> it: 26, PLL:-476.819
#> it: 27, PLL:-475.002
#> it: 28, PLL:-473.644
#> it: 29, PLL:-472.703
#> it: 30, PLL:-472.1
#> it: 31, PLL:-471.708
#> it: 32, PLL:-471.444
#> it: 33, PLL:-471.266
#> it: 34, PLL:-471.143
#> it: 35, PLL:-471.058
#> it: 36, PLL:-471.001
#> it: 37, PLL:-470.963
#> it: 38, PLL:-470.935
#> it: 39, PLL:-470.914
#> it: 40, PLL:-470.897
#> it: 41, PLL:-470.883
#> it: 42, PLL:-470.871
#> it: 43, PLL:-470.86
#> it: 44, PLL:-470.851
#> it: 45, PLL:-470.842
#> it: 46, PLL:-470.833
#> it: 47, PLL:-470.824
#> it: 48, PLL:-470.814
#> it: 49, PLL:-470.803
#> it: 50, PLL:-470.789
#> it: 51, PLL:-470.771
#> it: 52, PLL:-470.75
#> it: 53, PLL:-470.73
#> it: 54, PLL:-470.71
#> it: 55, PLL:-470.692
#> it: 56, PLL:-470.674
#> it: 57, PLL:-470.656
#> it: 58, PLL:-470.638
#> it: 59, PLL:-470.621
#> it: 60, PLL:-470.603
#> it: 61, PLL:-470.585
#> it: 62, PLL:-470.569
#> it: 63, PLL:-470.555
#> it: 64, PLL:-470.542
#> it: 65, PLL:-470.53
#> it: 66, PLL:-470.52
#> it: 67, PLL:-470.511
#> it: 68, PLL:-470.5
#> it: 69, PLL:-470.492
#> it: 70, PLL:-470.481
#> it: 71, PLL:-470.471
#> it: 72, PLL:-470.47
#> it: 73, PLL:-470.47
#> it: 74, PLL:-470.47
#> it: 75, PLL:-470.47
#> it: 76, PLL:-470.469
#> it: 77, PLL:-470.469
#> it: 78, PLL:-470.469